requirements.txt
nextcord~=2.0.0a10
python-dotenv~=0.15.0